Scan and parse regex
To scan regular expressions, we need to identify when we are looking for an infix token explicitly, at least that's how JSONata does it in Javascript, so that's what I did here.

This includes some refactoring to the tokenizer and parser to add little more clarity.

Regular expressions don't do anything yet, they are just now parsed as part of the AST.