Tokenize numbers with str.parse instead of json::Number
The json::Number situation is not good, we want to add some more fidelity to our number computations and do them in integer if possible, however with json::Number everything ends up in float.

This will make evaluation more complicated as we have to deal with types, but the rules should be simple enough.